+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOTenantSettings/MSFT_SPOTenantSettings.psm1
@@ -14,11 +14,6 @@ function Get-TargetResource
         $EnableAzureADB2BIntegration,
 
         [Parameter()]
-        [ValidateSet('ExternalUserAndGuestSharing', 'Disabled', 'ExternalUserSharingOnly', 'ExistingExternalUserSharingOnly')]
-        [System.String]
-        $OneDriveSharingCapability,
-
-        [Parameter()]
         [System.UInt32]
         $MinCompatibilityLevel,
 
@@ -219,8 +214,7 @@ function Get-TargetResource
             'DenySelectSGsInODBListInTenant',
             'DenySelectSecurityGroupsInSPSitesList',
             'AllowSelectSecurityGroupsInSPSitesList',
-            'EnableAzureADB2BIntegration',
-            'OneDriveSharingCapability')
+            'EnableAzureADB2BIntegration')
 
         $response = Invoke-PnPSPRestMethod -Method Get `
             -Url "$((Get-MSCloudLoginConnectionProfile -Workload PnP).AdminUrl)/_api/SPO.Tenant?`$select=$($parametersToRetrieve -join ',')"
@@ -234,7 +228,6 @@ function Get-TargetResource
             DenySelectSecurityGroupsInSPSitesList                  = $response.DenySelectSecurityGroupsInSPSitesList
             AllowSelectSecurityGroupsInSPSitesList                 = $response.AllowSelectSecurityGroupsInSPSitesList
             EnableAzureADB2BIntegration                            = $response.EnableAzureADB2BIntegration
-            OneDriveSharingCapability                              = $response.OneDriveSharingCapability
             MinCompatibilityLevel                                  = $MinCompat
             MaxCompatibilityLevel                                  = $MaxCompat
             SearchResolveExactEmailOrUPN                           = $SPOTenantSettings.SearchResolveExactEmailOrUPN
@@ -299,11 +292,6 @@ function Set-TargetResource
         [Parameter()]
         [System.Boolean]
         $EnableAzureADB2BIntegration,
-
-        [Parameter()]
-        [ValidateSet('ExternalUserAndGuestSharing', 'Disabled', 'ExternalUserSharingOnly', 'ExistingExternalUserSharingOnly')]
-        [System.String]
-        $OneDriveSharingCapability,
 
         [Parameter()]
         [System.UInt32]
@@ -499,7 +487,6 @@ function Set-TargetResource
     $CurrentParameters.Remove('DenySelectSecurityGroupsInSPSitesList') | Out-Null
     $CurrentParameters.Remove('AllowSelectSecurityGroupsInSPSitesList') | Out-Null
     $CurrentParameters.Remove('EnableAzureADB2BIntegration') | Out-Null
-    $CurrentParameters.Remove('OneDriveSharingCapability') | Out-Null
 
     $CurrentParameters.Remove('TenantDefaultTimezone') | Out-Null # this one is updated separately using Graph
     if ($CurrentParameters.Keys.Contains('UserVoiceForFeedbackEnabled'))
@@ -562,12 +549,6 @@ function Set-TargetResource
             $paramsToUpdate.Add('EnableAzureADB2BIntegration', $EnableAzureADB2BIntegration)
         }
 
-        if ($null -ne $OneDriveSharingCapability)
-        {
-            $needToUpdate = $true
-            $paramsToUpdate.Add('OneDriveSharingCapability', $OneDriveSharingCapability)
-        }
-
         if ($needToUpdate)
         {
             Write-Verbose -Message 'Updating properties via REST PATCH call.'
@@ -603,11 +584,6 @@ function Test-TargetResource
         [Parameter()]
         [System.Boolean]
         $EnableAzureADB2BIntegration,
-
-        [Parameter()]
-        [ValidateSet('ExternalUserAndGuestSharing', 'Disabled', 'ExternalUserSharingOnly', 'ExistingExternalUserSharingOnly')]
-        [System.String]
-        $OneDriveSharingCapability,
 
         [Parameter()]
         [System.UInt32]